+++ dependency/streamline/include/streamline/inspector.hpp
@@ -9,8 +9,9 @@ class Inspector {
 public:
     using DataType = T;
 
-    Inspector()                                         = default;
-    virtual ~Inspector()                                = default;
-    virtual void inspect(System&, DataType const& data) = 0;
+    Inspector()                                                       = default;
+    virtual ~Inspector()                                              = default;
+    virtual void inspect(System&, DataType const& data, uint64_t tag) = 0;
+    virtual bool accept(System&, uint64_t) { return true; }
 };
 }  // namespace streamline