+++ examples/client/processor/control.cpp
@@ -7,18 +7,22 @@
 LOGLET_MODULE2(p, ctrl);
 #define LOGLET_CURRENT_MODULE &LOGLET_MODULE_REF2(p, ctrl)
 
-void CtrlPrint::inspect(streamline::System&, DataType const& message) NOEXCEPT {
+void CtrlPrint::inspect(streamline::System&, DataType const& message, uint64_t) NOEXCEPT {
     VSCOPE_FUNCTION();
     message->print();
 }
 
-void CtrlOutput::inspect(streamline::System&, DataType const& message) NOEXCEPT {
+void CtrlOutput::inspect(streamline::System&, DataType const& message, uint64_t tag) NOEXCEPT {
     VSCOPE_FUNCTION();
     auto payload = message->payload();
     auto data    = reinterpret_cast<uint8_t const*>(payload.data());
     auto size    = payload.size();
     for (auto const& output : mConfig.outputs) {
         if (!output.ctrl_support()) continue;
+        if(!output.accept_tag(tag)) {
+            XDEBUGF(OUTPUT_PRINT_MODULE, "ctrl: tag %llX not accepted", tag);
+            continue;
+        }
         if (output.print) {
             XINFOF(OUTPUT_PRINT_MODULE, "ctrl: %zd bytes", size);
         }
@@ -26,7 +30,7 @@ void CtrlOutput::inspect(streamline::System&, DataType const& message) NOEXCEPT 
     }
 }
 
-void CtrlEvents::inspect(streamline::System&, DataType const& message) NOEXCEPT {
+void CtrlEvents::inspect(streamline::System&, DataType const& message, uint64_t) NOEXCEPT {
     VSCOPE_FUNCTION();
     if (on_cell_id) {
         auto cell_id = dynamic_cast<format::ctrl::CellId*>(message.get());