+++ examples/client/processor/lpp.hpp
@@ -18,7 +18,7 @@ class LppXerOutput : public streamline::Inspector<lpp::Message> {
 public:
     LppXerOutput(OutputConfig const& output) : mOutput(output) {}
 
-    void inspect(streamline::System&, DataType const& message) NOEXCEPT override;
+    void inspect(streamline::System&, DataType const& message, uint64_t tag) NOEXCEPT override;
 
 private:
     OutputConfig const& mOutput;
@@ -28,7 +28,7 @@ class LppUperOutput : public streamline::Inspector<lpp::Message> {
 public:
     LppUperOutput(OutputConfig const& output) : mOutput(output) {}
 
-    void inspect(streamline::System&, DataType const& message) NOEXCEPT override;
+    void inspect(streamline::System&, DataType const& message, uint64_t tag) NOEXCEPT override;
 
 private:
     OutputConfig const& mOutput;