+++ examples/client/processor/lpp2rtcm.hpp
@@ -14,7 +14,7 @@ public:
     Lpp2Rtcm(OutputConfig const& output, Lpp2RtcmConfig const& config);
     ~Lpp2Rtcm() override;
 
-    void inspect(streamline::System&, DataType const& message) override;
+    void inspect(streamline::System&, DataType const& message, uint64_t tag) override;
 
 private:
     std::unique_ptr<generator::rtcm::Generator> mGenerator;