+++ examples/client/processor/lpp2spartn.hpp
@@ -14,7 +14,7 @@ public:
     Lpp2Spartn(OutputConfig const& output, Lpp2SpartnConfig const& config);
     ~Lpp2Spartn() override;
 
-    void inspect(streamline::System&, DataType const& message) override;
+    void inspect(streamline::System&, DataType const& message, uint64_t tag) override;
 
 private:
     std::unique_ptr<generator::spartn::Generator> mGenerator;