+++ examples/client/processor/tokoro.cpp
@@ -88,7 +88,7 @@ void TokoroEphemerisUbx::handle_bds(format::ubx::RxmSfrbx* sfrbx) {
     }
 }
 
-void TokoroEphemerisUbx::inspect(streamline::System&, DataType const& message) {
+void TokoroEphemerisUbx::inspect(streamline::System&, DataType const& message, uint64_t) {
     VSCOPE_FUNCTION();
     auto ptr = message.get();
     if (!ptr) return;
@@ -109,7 +109,7 @@ void TokoroEphemerisUbx::inspect(streamline::System&, DataType const& message) {
 //
 //
 
-void TokoroLocation::inspect(streamline::System&, DataType const& location) {
+void TokoroLocation::inspect(streamline::System&, DataType const& location, uint64_t) {
     VSCOPE_FUNCTION();
     mTokoro.update_location_information(location);
 }
@@ -125,6 +125,7 @@ Tokoro::Tokoro(OutputConfig const& output, TokoroConfig const& config,
     mGenerator = std::unique_ptr<generator::tokoro::Generator>(new generator::tokoro::Generator{});
     mReferenceStation = nullptr;
     mPeriodicTask     = nullptr;
+    mOutputTag        = 0;
 
     mGenerator->set_iod_consistency_check(mConfig.iod_consistency_check);
     mGenerator->set_rtoc(mConfig.rtoc);
@@ -360,6 +361,10 @@ void Tokoro::generate(ts::Tai const& generation_time) {
         // TODO(ewasjon): These message should be passed back into the system
         for (auto const& output : mOutput.outputs) {
             if (!output.rtcm_support()) continue;
+            if (!output.accept_tag(mOutputTag)) {
+                XDEBUGF(OUTPUT_PRINT_MODULE, "tag %llX not accepted", mOutputTag);
+                continue;
+            }
             if (output.print) {
                 XINFOF(OUTPUT_PRINT_MODULE, "rtcm: %04d (%zd bytes)", submessage.id(), size);
             }
@@ -369,7 +374,7 @@ void Tokoro::generate(ts::Tai const& generation_time) {
     }
 }
 
-void Tokoro::inspect(streamline::System&, DataType const& message) {
+void Tokoro::inspect(streamline::System&, DataType const& message, uint64_t) {
     VSCOPE_FUNCTION();
     ASSERT(mGenerator, "generator is null");