+++ examples/client/processor/tokoro.hpp
@@ -11,12 +11,12 @@
 #include <format/nav/gal/inav.hpp>
 #include <format/nav/gps/lnav.hpp>
 #include <format/ubx/messages/rxm_sfrbx.hpp>
+#include <generator/tokoro/constant.hpp>
 #include <generator/tokoro/coordinate.hpp>
 #include <generator/tokoro/generator.hpp>
-#include <generator/tokoro/constant.hpp>
 #include <lpp/location_information.hpp>
-#include <scheduler/scheduler.hpp>
 #include <scheduler/periodic.hpp>
+#include <scheduler/scheduler.hpp>
 
 #include "config.hpp"
 #include "lpp.hpp"
@@ -33,7 +33,7 @@ public:
     void process_ephemeris(ephemeris::GalEphemeris const& ephemeris) NOEXCEPT;
     void process_ephemeris(ephemeris::BdsEphemeris const& ephemeris) NOEXCEPT;
 
-    void inspect(streamline::System&, DataType const& message) override;
+    void inspect(streamline::System&, DataType const& message, uint64_t tag) override;
 
     void vrs_mode_fixed();
     void vrs_mode_dynamic();
@@ -52,6 +52,7 @@ private:
     std::unique_ptr<generator::tokoro::Generator>        mGenerator;
     std::shared_ptr<generator::tokoro::ReferenceStation> mReferenceStation;
     std::unique_ptr<scheduler::PeriodicTask>             mPeriodicTask;
+    uint64_t                                             mOutputTag;
 };
 
 class TokoroEphemerisUbx : public streamline::Inspector<UbxMessage> {
@@ -67,7 +68,7 @@ public:
     void handle_bds_d1(format::ubx::RxmSfrbx* sfrbx);
     void handle_bds(format::ubx::RxmSfrbx* sfrbx);
 
-    void inspect(streamline::System&, DataType const& message) override;
+    void inspect(streamline::System&, DataType const& message, uint64_t tag) override;
 
 private:
     Tokoro&                                    mTokoro;
@@ -80,7 +81,7 @@ class TokoroLocation : public streamline::Inspector<lpp::LocationInformation> {
 public:
     TokoroLocation(Tokoro& tokoro) : mTokoro(tokoro) {}
 
-    void inspect(streamline::System&, DataType const& location) override;
+    void inspect(streamline::System&, DataType const& location, uint64_t tag) override;
 
 private:
     Tokoro& mTokoro;